server/requirements.txt
arrow==0.10.0
blessed==1.14.2
certifi==2017.7.27.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.5.0
Django==1.11.6
django-picklefield==1.0.0
django-q==0.8.1
django-rest-swagger==2.1.2
djangorestframework==3.7.0
facebook-sdk==2.0.0
idna==2.6
itypes==1.1.0
Jinja2==2.9.6
MarkupSafe==1.0
oauthlib==2.0.4
openapi-codec==1.3.2
pkg-resources==0.0.0
PyJWT==1.5.3
python-dateutil==2.6.1
python3-openid==3.1.0
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
simplejson==3.11.1
six==1.11.0
social-auth-app-django==1.2.0
social-auth-core==1.4.0
uritemplate==3.0.0
urllib3==1.22
wcwidth==0.1.7